+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - RandomElement extension method for IList.
+- (Editor) Option to clear active selection on play.
 
 ## [1.0.0] - 2023-01-16